backend/requirements.txt
flask[async]
flask-cors
waitress
python-dotenv
pydantic
pydantic-settings
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-pinecone
pinecone
pymupdf
gdown
psutil
unzip
